Split handler tests into separate files
